fix: use dark theme for onboarding screen
