--- lib/theme_colors.dart
@@ -1,10 +1,0 @@
-import 'dart:ui';
-
-class ThemeColors {
-  static const accentColor = Color(0xFF009FDF);
-  static const detailColor = Color(0xFF121312);
-  static const textFieldColor = Color(0xFF1D1D1F);
-
-  static const lightAccentColor = Color(0xFF006591);
-  static const lightTextFieldColor = Color(0xFFEDEDEF);
-}
